# src/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InkyPi API",
    "description": "HTTP API for the InkyPi e-ink display controller. Manage playlists, plugins, display refresh, settings, and system health.",
    "version": "0.60.6",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local InkyPi device (default dev port)"
    },
    {
      "url": "http://inkypi.local",
      "description": "InkyPi device on LAN via mDNS"
    }
  ],
  "paths": {
...
    "/api/current_image": {
      "get": {
        "tags": [
          "Display"
        ],
        "summary": "Current display image",
        "description": "Returns the most recently rendered image as PNG. Supports conditional GET via If-Modified-Since for efficient polling.",
        "operationId": "getCurrentImage",
        "parameters": [
          {
            "name": "If-Modified-Since",
            "in": "header",
            "description": "Return 304 Not Modified if image has not changed since this timestamp",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image of the current display",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified \u2014 image unchanged since If-Modified-Since"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/display-next": {
      "post": {
        "tags": [
          "Display"
        ],
        "summary": "Advance to next plugin in playlist",
        "description": "Triggers an immediate display refresh, advancing to the next eligible plugin in the active playlist. Rate-limited to prevent rapid successive calls.",
        "operationId": "displayNext",
        "responses": {
          "200": {
            "description": "Refresh triggered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayNextResponse"
                },
                "example": {
                  "success": true,
                  "plugin_id": "clock",
                  "plugin_instance": "My Clock"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "An error occurred"
          }
        }
      },
...
      "DisplayNextResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "plugin_id": {
            "type": "string",
            "description": "Plugin type that was rendered",
            "example": "clock"
          },
          "plugin_instance": {
            "type": "string",
            "description": "Plugin instance name that was rendered",
            "example": "My Clock"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request \u2014 invalid parameters or missing required fields",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded \u2014 try again after a short delay",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
...
}